+++ src/cavalry.cpp
@@ -53,7 +53,7 @@
             }
         }
         else {
-            for (int i=0; i<v.size(); i++) {
+            for (int i=0; i<static_cast<int>(v.size()); i++) {
                 if (v[i]->getY()-2<=this->y+this->mv_speed+2 && v[i]->getY()-1>this->y && v[i]->getX()==this->x) {
                     this->setspeed(1);
                     v[i]->setspeed(1);
